+++ src/caetgory.py
@@ -1,24 +1,4 @@
-# Class Product
-class Product:
-    """Accepts 4 required properties and returns data using 4 methods:
-    name - returns the name
-    description - returns the product description
-    price - returns the product price
-    quantity - returns the product quantity
-    """
-
-    name: str
-    description: str
-    price: int
-    quantity: int
-
-    def __init__(self, name, description, price, quantity):
-        self.name = name
-        self.description = description
-        self.price = price
-        self.quantity = quantity
-
-
+# Class 
 class Category:
     """Accepts 3 required properties and returns data using 3 methods
     name - returns category name
@@ -39,6 +19,13 @@ class Category:
     def __init__(self, name, description, products):
         self.name = name
         self.description = description
-        self.products = products
+        self.__products = products
         Category.category_count += 1
         Category.product_count += len(products)
+
+    def add_product(self, new_item=[]):
+        self.__products = [new_item]
+    
+    @property.getter
+    def products(self):
+        return self.__products
