+++ src/product.py
@@ -1,0 +1,24 @@
+# Class Product
+class Product:
+    """Accepts 4 required properties and returns data using 4 methods:
+    name - returns the name
+    description - returns the product description
+    price - returns the product price
+    quantity - returns the product quantity
+    """
+
+    name: str
+    description: str
+    price: int
+    quantity: int
+
+    def __init__(self, name, description, price, quantity):
+        self.name = name
+        self.description = description
+        self.price = price
+        self.quantity = quantity
+
+    def new_product(self, update={}):
+        for i, key in enumerate(update):
+            self.key = update[key]
+            print(self.key)
